Replace most u32/usize/u64 in blockstore and blobstore interfaces with byte_unit::Byte

+++ crates/blobstore/src/interface.rs
@@ -1,5 +1,6 @@
 use anyhow::Result;
 use async_trait::async_trait;
+use byte_unit::Byte;
 use futures::stream::BoxStream;
 use std::fmt::Debug;
 
@@ -46,7 +47,7 @@ pub trait BlobStore {
     async fn num_nodes(&self) -> Result<u64>;
     fn estimate_space_for_num_blocks_left(&self) -> Result<u64>;
     // virtual means "space we can use" as opposed to "space it takes on the disk" (i.e. virtual is without headers, checksums, ...)
-    fn virtual_block_size_bytes(&self) -> u32;
+    fn virtual_block_size_bytes(&self) -> Byte;
 
     // TODO load_block_depth is only needed for our c++ bindings of the stats tool. Remove them.
     async fn load_block_depth(&self, _id: &cryfs_blockstore::BlockId) -> Result<Option<u8>>;
